feat: implement Pinia chat store with message handling, room switching, and online user tracking
